added more flow types #1384
