+++ app/assets/javascripts/admin/views/task/task_list_view.js
@@ -9,11 +9,11 @@ import Toast from "libs/toast";
 import AnonymousTaskLinkModal from "admin/views/task/anonymous_task_link_modal";
 import TaskSearchForm from "admin/views/task/task_search_form";
 import messages from "messages";
-import { deleteTask } from "admin/admin_rest_api";
+import { deleteTask, getTasksByQuery } from "admin/admin_rest_api";
 import TemplateHelpers from "libs/template_helpers";
 import FormatUtils from "libs/format_utils";
 import type { APITaskType, APITaskTypeType } from "admin/api_flow_types";
-import Request from "libs/request";
+import type { QueryObjectType } from "admin/views/task/task_search_form";
 
 const { Column } = Table;
 const { Search } = Input;
@@ -33,42 +33,9 @@ class TaskListView extends React.PureComponent<Props, State> {
     searchQuery: "",
   };
 
-  componentDidMount() {
-    const queryObject = {
-      _id: { $oid: "598d8780cd682fff03a3b26a" },
-    };
-    this.fetchData(queryObject);
-  }
-
-  async fetchData(queryObject: Object) {
-    const responses = await Request.sendJSONReceiveJSON("/api/queries", {
-      params: { type: "task" },
-      data: queryObject,
-    });
-
-    const tasks = responses.map(response => {
-      // apply some defaults
-      response.type = {
-        summary: Utils.__guard__(response.type, x => x.summary) || "<deleted>",
-        id: Utils.__guard__(response.type, x1 => x1.id) || "",
-      };
-
-      if (response.tracingTime == null) {
-        response.tracingTime = 0;
-      }
-      // convert bounding box
-      if (response.boundingBox != null) {
-        const { topLeft, width, height, depth } = response.boundingBox;
-        response.boundingBox = topLeft.concat([width, height, depth]);
-      } else {
-        response.boundingBox = [];
-      }
-
-      return response;
-    });
-
+  async fetchData(queryObject: QueryObjectType) {
     this.setState({
-      tasks,
+      tasks: await getTasksByQuery(queryObject),
     });
   }
 
@@ -125,7 +92,7 @@ class TaskListView extends React.PureComponent<Props, State> {
             <Table
               dataSource={Utils.filterWithSearchQueryAND(
                 this.state.tasks,
-                ["team", "projectName", "id", "dataSet", "created", "type", "experience"],
+                ["team", "projectName", "id", "dataSet", "created", "type", "neededExperience"],
                 this.state.searchQuery,
               )}
               rowKey="id"
@@ -166,7 +133,7 @@ class TaskListView extends React.PureComponent<Props, State> {
                 render={(__, task: APITaskType) => (
                   <div className="nowrap">
                     {TemplateHelpers.formatTuple(task.editPosition)} <br />
-                    <span>{TemplateHelpers.formatTuple(task.boundingBox)}</span>
+                    <span>{TemplateHelpers.formatTuple(task.boundingBoxVec6)}</span>
                   </div>
                 )}
               />
